kobo fix

+++ src/main/java/com/eazybrew/vend/paystack/dto/request/InitializeTransactionRequest.java
@@ -5,14 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class InitializeTransactionRequest {
     private String email;
-    private BigDecimal amount;
+    private Long amount;
 }
